+++ src/rich_gradient/pretty.py
@@ -10,7 +10,12 @@ from rich.console import Console, JustifyMethod, OverflowMethod
 from rich.highlighter import Highlighter
 from rich.pretty import Pretty as RichPretty
 
-from rich_gradient.gradient import ColorType, Gradient, HighlightRegexType, HighlightWordsType
+from rich_gradient.gradient import (
+    ColorType,
+    Gradient,
+    HighlightRegexType,
+    HighlightWordsType,
+)
 
 
 class Pretty(Gradient):
@@ -59,17 +64,17 @@ class Pretty(Gradient):
         expand_all: bool = False,
         margin: int = 0,
         insert_line: bool = False,
-        colors: Optional[Sequence[ColorType]] = None,
-        bg_colors: Optional[Sequence[ColorType]] = None,
+        colors: Sequence[ColorType] | None = None,
+        bg_colors: Sequence[ColorType] | None = None,
         rainbow: bool = False,
         hues: int = 5,
         repeat_scale: float = 2.0,
         expand: bool = True,
         justify: AlignMethod = "left",
         vertical_justify: VerticalAlignMethod = "middle",
-        console: Optional[Console] = None,
-        highlight_words: Optional[HighlightWordsType] = None,
-        highlight_regex: Optional[HighlightRegexType] = None,
+        console: Console | None = None,
+        highlight_words: HighlightWordsType | None = None,
+        highlight_regex: HighlightRegexType | None = None,
     ) -> None:
         """Initialize a gradient-enabled Rich pretty-printer."""
         pretty = RichPretty(